Removes test script's use of a loop device
Removes the test script's use of an intermediary loop block device (/dev/loop6)
for an empty file to temporarily associated with while the filesystem is
created on it.

There generally no need to setup a loop device using 'losetup' to do this: mkfs
and partclone operate on files with identical behavior. The only exception is
mkfs.ntfs unless the --force argument is used.

Because of this, the call to mkfs.$FS has been abstracted to handle this.