Add helper script to create release tarball
Adds a small script to create a gzipped tarball including all the files in
current working directory, except the script. Prior releases of partclone-utils
have contained a subfolder within the tarball, and users [1] have called for
this to remain. Therefore the tarball command uses sed filename transform to
ensure the subfolder approach is remains [2].

[1] https://sourceforge.net/p/partclone-utils/discussion/general/thread/5f07414b/#52ac/665e/97c4/6b9a

[2] https://sourceforge.net/p/partclone-utils/discussion/general/thread/5f07414b/#ae91