formatting: add clang's configuration file
clang allows to define a configuration that allow to get a consistent
format across all C files. We can even install an extension in VS Code
to automatically format the file when we save.

The main big change in the formatting is the use of spaces only to indent
the code. The previous format was using tabs(8) and chars(4) for indenting,
which is not easy to follow with most editors.