partclone_imageinfo: fix block count verification
The code was not handling v2 image correctly. It was specific to v1 image.
In particular, it was assuming a CRC32 was present after every block.

Since partclone 0.3.0, the image format is v2 and the default options for
checksum are:
- checksum algo: -a1 (CRC32)
- blocks per checksum: -k256
- checksum reseed: yes

This commit updates the code to correctly handle all use cases.

Signed-off-by: Patrick Rouleau <pfrouleau@gmail.com>