formatting: remove parenthesis from return statement
I do not remember exactly what was the reasoning for using parenthesis
with a return statement, but it was something related to the old
compilers (like <'90s). They are removed to follow more closely partclone's
coding style.

NB: I used GreatCode with "-code_remove_return_paren-" to partially
automate the task, then only staged the modification related to return.